moved camera scripts folder
