+++ Assets/_Application/Cameras/Compass.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.UI;
 
-namespace Netherlands3D.Twin.PackageStagingArea.eu.netherlands3d.cameras.Runtime.Scripts.Cameras
+namespace Netherlands3D.Twin.Cameras
 {
     public class Compass : MonoBehaviour
     {
